+++ src/pages/characterShare/CharacterSharePage.tsx
@@ -2,8 +2,7 @@ import { useRef } from 'react';
 import { toPng } from 'html-to-image';
 import saveAs from 'file-saver';
 import { showToast } from '@/shared/ui/Toast';
-import Button from '@/shared/ui/Button';
-import { TextButton } from '@/shared/design-system/ui';
+import { Button, TextButton } from '@/shared/design-system/ui';
 import { useLoaderData, useNavigate } from 'react-router';
 import { useTheme } from 'styled-components';
 import { ArrowLeft, Download } from '@/shared/assets/svgs/icon';
