+++ src/pages/expenseDetail/ExpenseDetailPage.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo, useState } from 'react';
 import { useLoaderData, useNavigate } from 'react-router';
 import { useTheme } from 'styled-components';
 import { ArrowLeft } from '@/shared/assets/svgs/icon';
-import Button from '@/shared/ui/Button';
 import ButtonGroup from '@/shared/ui/ButtonGroup';
+import { Button } from '@/shared/design-system/ui';
 import Header from '@/shared/ui/Header';
 import Text from '@/shared/ui/Text';
 import Flex from '@/shared/ui/Flex';
