+++ src/pages/groupNameSetup/GroupNameSetupPage.tsx
@@ -8,7 +8,7 @@ import Text from '@/shared/ui/Text';
 import Header from '@/shared/ui/Header';
 import DescriptionField from '@/shared/ui/DescriptionField';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
-import Button from '@/shared/ui/Button';
+import { Button } from '@/shared/design-system/ui';
 import Input from '@/shared/ui/Input';
 import * as S from './GroupNameSetupPage.styles';
