--- src/shared/ui/Button/index.styles.ts
@@ -1,60 +1,0 @@
-import styled, { css } from 'styled-components';
-import { TextVariant } from '@/shared/ui/Text/index.styles';
-import { ButtonSize, ButtonVariant } from './index.type';
-
-interface StyledButtonProps {
-  $variant: ButtonVariant;
-  $size: ButtonSize;
-}
-
-const buttonVariants = {
-  primary: css`
-    background: ${({ theme }) => theme.color.semantic.orange.default};
-    color: ${({ theme }) => theme.color.semantic.text.inverse};
-  `,
-  secondary: css`
-    background: ${({ theme }) => theme.color.semantic.secondary.default};
-    color: ${({ theme }) => theme.color.semantic.text.strong};
-  `,
-  tertiary: css`
-    background: transparent;
-    color: ${({ theme }) => theme.color.semantic.text.strong};
-    border: 1px solid ${({ theme }) => theme.color.semantic.border.default};
-  `,
-  text: css`
-    background: transparent;
-    padding: 0;
-    color: ${({ theme }) => theme.color.semantic.text.default};
-    ${TextVariant('body1R')};
-  `,
-};
-
-const buttonSizes = {
-  md: css`
-    padding: ${({ theme }) => `${theme.unit[12]} ${theme.unit[16]}`};
-    ${TextVariant('body1Sb')};
-  `,
-  sm: css`
-    padding: ${({ theme }) => `${theme.unit[8]} ${theme.unit[16]}`};
-    ${TextVariant('body2Sb')};
-  `,
-};
-
-export const Button = styled.button<StyledButtonProps>`
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  gap: ${({ theme }) => theme.unit[4]};
-  border-radius: ${({ theme }) => theme.radius.circle};
-  cursor: pointer;
-  border: none;
-  white-space: nowrap;
-  ${({ $size }) => buttonSizes[$size]};
-  ${({ $variant }) => buttonVariants[$variant]};
-  &:disabled {
-    cursor: not-allowed;
-    background: ${({ theme }) =>
-      theme.color.semantic.background.normal.disabled};
-    color: ${({ theme }) => theme.color.semantic.text.inverse};
-  }
-`;
